// backups/release-candidates/zenova-definitiva-2026-03-01_1437/types.ts
export interface User {
  id: string;
  name: string;
  email: string;
  password: string;
  role: 'admin' | 'viewer' | 'nurse';
  mustChangePassword?: boolean;  // camelCase en TypeScript
  nurseId?: string | null;       // camelCase en TypeScript
}